split simulation file into simulation module
