+++ bsb/simulation/results.py
@@ -1,0 +1,3 @@
+class SimulationResult:
+    def __init__(self):
+        pass
